mod: change return type from long to int

+++ src/main/java/model/Usuario.java
@@ -54,7 +54,7 @@ public class Usuario implements Serializable {
         setDir(dir);
     }
 
-    public long getClienteId() {
+    public int getClienteId() {
         return clienteId;
     }
